Added Mining Scow (Reg) and Mining Scow (PCA)
Behold, the "humble mining scow". I based it initially off the star barge just as a starting point. I basically stripped the entire statline down to the brass tacks. Mining ships ought to be fast, with gun ports and decent cargo, so in exchange for all that I lowered shields/hull/fuel capacity, made it not very heat efficient, and seriously limited weapons/engine space. It also has only one bunk to boot.

The Mining Scow (Reg) uses the new mining lasers, but everything else is already in the game. Mining Scow (PCA) also uses the Plenum Chamber Afterburner, and has slightly different stats because of that. I figured a new afterburner might not make it in, so I made up both versions to be safe.